Bugfix: FIX session wipeout by reducing data stored within session
Because of a strange bug with CookieStore, adding the store of the
`state` value from params wiped some keys on the redirect (`auth` and
`state`)

Don't know why, but switching session store solves the issue.
We can't store personal data on our side (legals related), so we can
just reduce the size of what we store from `auth` hash from OmniAuth.

Tried to fix the issue with a non-regression test, but did not manage to
fix it..

Manually tested with staging env from FranceConnect, it works.